qt6: port shortcuts, QAudioSink, QActionGroup includes, qsizetype, QRegExp/midRef
Continues the Qt6 port; Qt5 still builds.

- Cross-version key shortcuts via new util/KeySeq.h (Qt5 rejects
  'modifier | key', Qt6 deletes 'modifier + key'); routed MainWindow,
  ChannelFindTools, ExternalViewerTools shortcut combos through it.
- AudioPlayer: QAudioOutput->QAudioSink and QAudioDeviceInfo->
  QMediaDevices/QAudioDevice behind QT_VERSION guards (BmsAudioOutput
  typedef); QAudioFormat set via setSampleFormat on Qt6.
- Add explicit <QActionGroup> includes (Qt6 drops transitive include) in
  MainWindow, SequenceTools, ExternalViewerTools, SequenceView.
- MasterView: cast QList::size() (qsizetype) for std::min on Qt6.
- ExternalViewer: include <QRegExp>, QString::midRef -> mid.
- ChannelInfoView: sample bit-depth via bytesPerSample() on Qt6.
- ScalarRegion: include <climits> for INT_MIN/MAX.

Remaining for Qt6: the PCM decoder format descriptor in audio/WaveData.cpp
and audio/WaveStream.cpp (QAudioFormat sampleSize/sampleType, which Qt6
QAudioFormat cannot represent).

+++ src/MasterView.cpp
@@ -295,7 +295,7 @@ void MiniMapView::UpdateBuffer()
 	for (int y=0; y<height(); y++){
 		int t = y * sview->viewLength / height();
 		int t2 = (y+1) * sview->viewLength / height();
-		if (std::min(t2, rmsCacheOfTicks.size()) > t){
+		if (std::min(t2, (int)rmsCacheOfTicks.size()) > t){
 			if (t2 > rmsCacheOfTicks.size())
 				t2 = rmsCacheOfTicks.size();
 			if (t < 0)
@@ -588,7 +588,7 @@ void MasterLaneView::UpdateBackBuffer(const QRect &rect)
 	for (int y=rect.bottom()+my; y>=rect.top()-my; y--){
 		int t = sview->Y2Time(y);
 		int t2 = sview->Y2Time(y-1);
-		if (std::min(t2, mview->rmsCacheOfTicks.size()) > t){
+		if (std::min(t2, (int)mview->rmsCacheOfTicks.size()) > t){
 			if (t2 > mview->rmsCacheOfTicks.size())
 				t2 = mview->rmsCacheOfTicks.size();
 			if (t < 0)
